* subversion/libsvn_wc/wc_db.c   (assert_blob_columns_equal): Cast printf args to (long int) in order to     match the format specifiers and so avoid compiler warnings.     The correct way would be to use APR_SIZE_T_FMT but that makes     localization difficult.  This is only temporary diagnostic code so     either way will do.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1001669 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_wc/wc_db.c
@@ -433,7 +433,7 @@ assert_blob_columns_equal(svn_sqlite__stmt_t *stmt1,
     return svn_error_createf(SVN_ERR_WC_CORRUPT, NULL,
                              "Blob of length %ld in statement 1 differs "
                              "from blob of length %ld in statement 2",
-                             len1, len2);
+                             (long int)len1, (long int)len2);
   return SVN_NO_ERROR;
 }
